+++ sdk/data/aztables/models.go
@@ -4,6 +4,7 @@
 package aztables
 
 import (
+	"strings"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
@@ -348,4 +349,9 @@ func toGeneratedStatusType(g *generated.GeoReplicationStatusType) *GeoReplicatio
 		return to.Ptr(GeoReplicationStatusUnavailable)
 	}
 	return nil
+}
+
+func prepareKey(key string) string {
+	// escape any single-quotes
+	return strings.ReplaceAll(key, "'", "''")
 }